Refactor event loop to get two players redrawn and moving
